chore: move timer component back to ui folder
